+++ decidim-initiatives/app/models/decidim/initiative.rb
@@ -139,7 +139,7 @@ module Decidim
     #
     # RETURNS string
     delegate :banner_image, to: :type
-    delegate :document_number_authorization_handler, :promoting_committee_enabled?, to: :type
+    delegate :document_number_authorization_handler, :promoting_committee_enabled?, :custom_signature_end_date_enabled?, to: :type
     delegate :type, :scope, :scope_name, to: :scoped_type, allow_nil: true
 
     # PUBLIC
@@ -244,7 +244,7 @@ module Decidim
         published_at: Time.current,
         state: "published",
         signature_start_date: Date.current,
-        signature_end_date: Date.current + Decidim::Initiatives.default_signature_time_period_length
+        signature_end_date: signature_end_date || Date.current + Decidim::Initiatives.default_signature_time_period_length
       )
     end
 
@@ -452,6 +452,11 @@ module Decidim
 
     # Allow ransacker to search on an Enum Field
     ransacker :state, formatter: proc { |int| states[int] }
+      
+    ransacker :type_id do
+      Arel.sql("decidim_initiatives_type_scopes.decidim_initiatives_types_id")
+    end
+      
     # method for sort_link by number of supports
 
     # TODO: OSP PR #969: Find a way to sort by support_counts